+++ test/router.test.ts
@@ -1,6 +1,10 @@
 import { Router } from '../src';
 
 describe('Test "Router"', () => {
+  beforeAll(() => {
+    document.body.innerHTML = '<p class="page"></p><a href="#u"></a>';
+  });
+
   test('"Router.current" before "Router.to" must to have empty values in "path" and "view()"', () => {
     expect(Router.current.path).toEqual('');
     expect(Router.current.view()).toEqual('');
@@ -10,8 +14,8 @@ describe('Test "Router"', () => {
     expect(Router.to('/')).resolves.not.toThrow();
   });
 
-  test('"Router.to" must use global "container" if there is not local one', () => {
-    Router.container = '.page ';
+  test('"Router.to" must use global "container" if there is no local one', () => {
+    Router.configure({ baseContainer: '.page ' });
     Router.add([
       {
         path: '/',
@@ -40,11 +44,10 @@ describe('Test "Router"', () => {
   test('"Router.to" must update "container" element with children that returns from "Route.view()"', async () => {
     await Router.to('/test');
 
-    const container = document.querySelector<HTMLDivElement>('.page');
+    const container = document.querySelector('.page');
 
     if (container) {
-      const child = container.innerText;
-      expect(child).toMatch('Test');
+      expect(container.innerHTML).toMatch('Test');
     }
   });
 
@@ -53,11 +56,26 @@ describe('Test "Router"', () => {
   });
 
   test('Navigating to any links without setting state, does not throw error', () => {
-    document.body.innerHTML = '<a href="#u"></a>';
-
     const a = document.querySelector('a');
     if (a) {
       expect(() => a.click()).not.toThrow();
+    }
+  });
+
+  test('Prefix path must be checked internally and users can navigate to path without prefix', async () => {
+    Router.configure({ basePrefix: '/pre' });
+    Router.add({
+      path: '/pre/fix',
+      view() {
+        return 'Prefix page';
+      },
+    });
+
+    await Router.to('/fix');
+    const pageElement = document.body.querySelector('.page');
+
+    if (pageElement) {
+      expect(pageElement.innerHTML).toBe('Prefix page');
     }
   });
 });